+++ packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @avaprotocol/sdk-js
 
+## 2.3.8
+
+### Patch Changes
+
+- 36ac928: Replace the old value getScheduleList to getSchedulesList
+- Updated dependencies [36ac928]
+  - @avaprotocol/types@2.2.8
+
+## 2.3.7
+
+### Patch Changes
+
+- 31667f8: Fix CustomCodeLang to match enum value
+- Updated dependencies [31667f8]
+  - @avaprotocol/types@2.2.7
+
 ## 2.3.6
 
 ### Patch Changes